Update FoiAttachmentMaskJob
When attachments are missing we parse the raw email again. After this we
can reload the attachment and see if the instance still exists in the
database, if not then reload based on the url part number and the
display filename (this is the same as in `AttachmentController` when we
decide which attachment to render). Once reloaded we can mask in the
same job without queuing a new job.